+++ riscv-rt/CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
 
+### Changed
+
+- Using `post_init` from `riscv-macros` instead of `riscv-rt-macros`.
+
 ## v0.17.1 - 2026-01-13
 
 ### Fixed